move `pollOutgoingPayment` out as fn; remove cyclic deps issue
move pollOutgoingPayment out of WalletService as not related to this service
